IGNITE-14496 "configuration-api" maven module extracted.
