+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/NamedConfigurationTree.java
