+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/annotation/ConfigurationRoot.java
