+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/tree/NamedListView.java
