+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/validation/Immutable.java
