+++ parent/pom.xml
@@ -149,6 +149,12 @@
                 <artifactId>ignite-configuration-annotation-processor</artifactId>
                 <version>${project.version}</version>
             </dependency>
+
+            <dependency>
+                <groupId>org.apache.ignite</groupId>
+                <artifactId>ignite-configuration-api</artifactId>
+                <version>${project.version}</version>
+            </dependency>
             
             <dependency>
                 <groupId>org.apache.ignite</groupId>
